[wip] JPS tests: Fix marking cross-compiled files as dirty
Before this change files that are not belonging to this compilation can be marted as dirty.
This causes rebuilding of other compilation.
Wrong files can be marked from shared global lookups and becose of bug in filtering files to mark as dirty:
files only belongs to current module or it's dependencies should be marked. The implementation assumed that each file belonged to no more than one target (`buildRootIndex.findJavaRootDescriptor(context, file)`), which is not true in case of mpp cross compilation.

This logic is fixed by checking all source roots, and filtering only that one that is belonging to current chunk or it's dependencies.